Refactored to be in transaction package
